+++ src/components/forms/EmailVerificationSend.tsx
@@ -5,7 +5,7 @@ import {
 } from '../../hooks/useEmailVerification';
 import { useFormContext } from 'react-hook-form';
 import { SignUpForm } from '../../pages/auth/SignUp';
-import Button from '../../widgets/Button';
+import ContainedButton from '../../widgets/buttons/ContainedButton';
 
 const EmailVerificationSend = ({
   state,
@@ -68,7 +68,7 @@ const EmailVerificationSend = ({
         })}
       />
       {errors.username && <span>{errors.username.message}</span>}
-      <Button
+      <ContainedButton
         type="button"
         onClick={onClick}
         disabled={
@@ -79,7 +79,7 @@ const EmailVerificationSend = ({
         }
       >
         인증메일 발송
-      </Button>
+      </ContainedButton>
     </>
   );
 };
