+++ src/pages/auth/SignIn.tsx
@@ -7,10 +7,11 @@ import { useSignInMutation } from '../../store/services/authApi';
 import { useAppSelector } from '../../hooks/rtk-hooks';
 import { RootState } from '../../store';
 import { useGoogleRecaptcha } from '../../hooks/useGoogleRecaptcha';
-import Button from '../../widgets/Button';
+import ItemButton from '../../widgets/buttons/ItemButton';
 import { GoSync } from 'react-icons/go';
 import { MdLogin } from 'react-icons/md';
-import Badge from '../../widgets/Badge';
+import ItemBadge from '../../widgets/badges/ItemBadge';
+import ContainedButton from '../../widgets/buttons/ContainedButton';
 
 export interface SignInForm {
   username: string;
@@ -120,7 +121,7 @@ const SignIn = () => {
         })}
       />
       {errors.password && <span>{errors.password.message}</span>}
-      <Button
+      <ContainedButton
         type="submit"
         disabled={isSubmitting}
         loading={isSubmitting}
@@ -128,13 +129,13 @@ const SignIn = () => {
         primary
       >
         {isSubmitting ? <GoSync className="animate-spin" /> : <MdLogin />}로그인
-      </Button>
-      <Badge primary pill className="text-xs">
+      </ContainedButton>
+      <ItemBadge primary pill className="text-xs">
         badge
-      </Badge>
-      <Button bounce warning rounded>
+      </ItemBadge>
+      <ItemButton bounce warning rounded>
         <MdLogin />
-      </Button>
+      </ItemButton>
       {reCaptchaElement}
     </form>
   );
