+++ src/pages/auth/SignUp.tsx
@@ -16,9 +16,9 @@ import useInterval from '../../hooks/useInterval';
 import EmailVerificationCode from '../../components/forms/EmailVerificationCode';
 import EmailVerificationSend from '../../components/forms/EmailVerificationSend';
 import PasswordConfirm from '../../components/forms/PasswordConfirm';
-import Button from '../../widgets/Button';
 import { GoSync } from 'react-icons/go';
 import { MdPersonAdd } from 'react-icons/md';
+import ContainedButton from '../../widgets/buttons/ContainedButton';
 
 export interface SignUpForm {
   username: string;
@@ -389,7 +389,7 @@ const SignUp = () => {
 
         <PasswordConfirm />
 
-        <Button
+        <ContainedButton
           type="submit"
           disabled={formMethods.formState.isSubmitting}
           loading={formMethods.formState.isSubmitting}
@@ -402,7 +402,7 @@ const SignUp = () => {
             <MdPersonAdd />
           )}
           회원가입
-        </Button>
+        </ContainedButton>
         {reCaptchaElement}
       </form>
     </FormProvider>
